Add containerized benchmark justfile target
This adds a total memory limit to make the comparison more apples to
apples.

However, at the moment it has exposed a bug in the configuration of some
of the database, such as RocksDB seems to be using much more memory than
intended. Once this is fixed the memory limit should be reduced to
something more appropriate to create some memory pressure in the
benchmark